+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from './fire.js';
 export * from './handler.js';
 export * from './utils.js';
+export * from './conninfo.js';
 
 // Re-export these utilities
 export type {