+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Extend env with client and server info
+- Implement ConnInfo helpers
 
 ## [0.3.1] - 2025-10-08