+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.3.2] - 2025-10-14
+
 ### Added
 
 - Extend env with client and server info
@@ -30,7 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial public release
 
-[unreleased]: https://github.com/fastly/hono-fastly-compute/compare/v0.3.1...HEAD
+[unreleased]: https://github.com/fastly/hono-fastly-compute/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/fastly/hono-fastly-compute/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/fastly/hono-fastly-compute/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/fastly/hono-fastly-compute/compare/v0.1.0...v0.3.0
 [0.1.0]: https://github.com/fastly/hono-fastly-compute/releases/tag/v0.1.0